requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
torch
torchaudio
transformers==4.36.2
scipy
numpy
pydantic==2.5.3
python-dotenv==1.0.0
soundfile==0.13.1
scikit-learn==1.3.2
joblib==1.3.2
librosa
audioread
google-generativeai>=0.8.0
groq
requests==2.31.0
